Add files via upload
As a data analyst by a Healthcare consultancy firm which has been conducting a survey on the state of global happiness annually. The World Happiness Report offers valuable insights into factors influencing happiness across countries. The firm wants to produce a report to find out whether there are demographic, regional, and/or economic characteristics that lead to a better life.